# local-shared-data/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/tts/synthesize": {
      "post": {
        "tags": [
          "tts",
          "tts"
        ],
        "summary": "Synthesize Speech",
        "description": "Generate speech from text using Hume.ai TTS API.\n\nArgs:\n    request: TTS request containing text, voice description, and format\n\nReturns:\n    Audio URL or base64-encoded audio data",
        "operationId": "synthesize_speech",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TTSRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TTSResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/ttimage/generateImage": {
      "post": {
        "tags": [
          "ttimage",
          "ttimage"
        ],
        "summary": "Generate Image",
        "description": "Generate an image using DALL-E based on a text prompt.\n\nArgs:\n    request: Dalle request containing the prompt and desired image size.\n\nReturns:\n    Image URL or image data",
        "operationId": "generate_image",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TTImageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TTImageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/lessons/create": {
      "post": {
        "tags": [
          "lessons",
          "lessons"
        ],
        "summary": "Create Lesson",
        "operationId": "create_lesson",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LessonCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/lessons/{lesson_id}/add_video": {
      "post": {
        "tags": [
          "lessons",
          "lessons"
        ],
        "summary": "Add Video To Lesson",
        "operationId": "add_video_to_lesson",
        "parameters": [
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Lesson Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LessonVideoAddResponse"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonVideoAddResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/lessons/{lesson_id}": {
      "get": {
        "tags": [
          "lessons",
          "lessons"
        ],
        "summary": "Get Lesson",
        "operationId": "get_lesson",
        "parameters": [
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Lesson Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/lessons/{lesson_id}/video/{index}": {
      "get": {
        "tags": [
          "lessons",
          "lessons"
        ],
        "summary": "Get Video By Index",
        "operationId": "get_video_by_index",
        "parameters": [
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Lesson Id"
            }
          },
          {
            "name": "index",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Index"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/lessons/{lesson_id}/video/{index}/has_next": {
      "get": {
        "tags": [
          "lessons",
          "lessons"
        ],
        "summary": "Has Next Video",
        "operationId": "has_next_video",
        "parameters": [
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Lesson Id"
            }
          },
          {
            "name": "index",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Index"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/scripts/generate-script-from-pdf": {
      "post": {
        "tags": [
          "Script Generation"
        ],
        "summary": "Generate Script From Pdf",
        "description": "Accepts a PDF upload and returns a ~2-minute video script with [BREAKPOINT] tags.",
        "operationId": "generate_script_from_pdf",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_Script_Generation-generate_script_from_pdf"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Body_Script_Generation-generate_script_from_pdf": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_Script Generation-generate_script_from_pdf"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LessonCreate": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Lesson title"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          }
        },
        "type": "object",
        "required": [
          "title",
          "user_id"
        ],
        "title": "LessonCreate"
      },
      "LessonRead": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Lesson title"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          }
        },
        "type": "object",
        "required": [
          "title",
          "id",
          "created_at",
          "user_id"
        ],
        "title": "LessonRead"
      },
      "LessonVideoAddResponse": {
        "properties": {
          "lesson_id": {
            "type": "string",
            "format": "uuid",
            "title": "Lesson Id"
          },
          "video_id": {
            "type": "string",
            "format": "uuid",
            "title": "Video Id"
          },
          "index": {
            "type": "integer",
            "title": "Index"
          }
        },
        "type": "object",
        "required": [
          "lesson_id",
          "video_id",
          "index"
        ],
        "title": "LessonVideoAddResponse"
      },
...
      "TTImageRequest": {
        "properties": {
          "prompt": {
            "type": "string",
            "title": "Prompt",
            "description": "Text prompt for image generation"
          },
          "n": {
            "type": "integer",
            "title": "N",
            "description": "Number of images to generate (max 1)"
          },
          "size": {
            "type": "string",
            "pattern": "^(1024x1024|512x512|256x256)$",
            "title": "Size",
            "description": "Image size: '1024x1024', '512x512', or '256x256'",
            "default": "1024x1024"
          }
        },
        "type": "object",
        "required": [
          "prompt",
          "n"
        ],
        "title": "TTImageRequest"
      },
      "TTImageResponse": {
        "properties": {
          "image_url": {
            "type": "string",
            "title": "Image Url"
          }
        },
        "type": "object",
        "required": [
          "image_url"
        ],
        "title": "TTImageResponse"
      },
      "TTSRequest": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text",
            "description": "Text to synthesize into speech"
          },
          "voice_description": {
            "type": "string",
            "title": "Voice Description",
            "description": "Description of the voice characteristics (e.g., 'warm female voice with slight accent')"
          },
          "format": {
            "type": "string",
            "pattern": "^(mp3|wav)$",
            "title": "Format",
            "description": "Audio format: 'mp3' or 'wav'",
            "default": "mp3"
          }
        },
        "type": "object",
        "required": [
          "text",
          "voice_description"
        ],
        "title": "TTSRequest"
      },
      "TTSResponse": {
        "properties": {
          "audio_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio Url"
          },
          "audio_base64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio Base64"
          },
          "format": {
            "type": "string",
            "title": "Format"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "format",
          "message"
        ],
        "title": "TTSResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/jwt/login"
          }
        }
      }
    }
  }
}